Updated CxImage source code to version 7.0.3 and fixed eMule to work with that (credits also go to eMule community)

+++ srchybrid/BaseClient.cpp
@@ -3304,7 +3304,7 @@ void CUpDownClient::SendPreviewAnswer(const CKnownFile* pForFile, CxImage** imgF
 			return;
 		}
 		BYTE* abyResultBuffer = NULL;
-		long nResultSize = 0;
+		int32_t nResultSize = 0;
 		if (!cur_frame->Encode(abyResultBuffer, nResultSize, CXIMAGE_FORMAT_PNG)){
 			ASSERT ( false );			
 			return;
